Rename DrawingToolRoot
