+++ packages/lib-classifier/src/store/feedback/helpers/get-feedback-from-tasks.js
@@ -2,7 +2,7 @@
 function getFeedbackFromTasks (tasks = {}) {
   const result = {}
   Object.entries(tasks).forEach(function getTaskRules ([taskKey, task]) {
-    const defaultFeedback = { enabled: false, rules: []}
+    const defaultFeedback = { enabled: false, rules: [] }
     const { feedback } = task
     const { enabled, rules } = feedback || defaultFeedback
     const taskHasFeedback =
